Fixed multi-polygon bug
Many multi-polygons may have a small, insignificant polygon as the first polygon,
and the old code relied on the first polygon as representing the main district, leading
to some strange behaviours. For example, clicking on Santa Barbara would first pan
to a small island in the Pacific before panning back to the click point for rendering
the popup. This fix supports rendering all polygons in the multi-polygon and will then
pan to the full region as expected.